drm hisi ade use correct stride to allow scanout and fb size to differ
The existing code here assumes that the framebuffer size is always the
scanout size.  There is no requirement that is true, and eg, if you start on
a 1080p mode and change mode in the framebuffer console to 720p later,
the framebuffer console remains at 1080p and the new mode is a subwindow on
the 1080p sized original framebuffer console.

This patch corrects the computations to use the right mix of information
about the memory layout (framebuffer stride) and the current scanout
(crtc active mode dimensions) so it works for all cases.

Signed-off-by: Andy Green <andy.green@linaro.org>