drm sysrq g mode hack
This patch let you force cycling between the remaining DRM modes from the
kernel alone.

It can only deal with one CRTC.

It's useful if a working display is needed for using or setting up a device,
but the HDMI compatibility is not 100%.  At least the user can check every
valid mode to see if it can work on his display.

Force the mode change by typing

Alt-Gr SysRq g

that is

 - hold down the right-hand ALT key
 - hold down the SysRq key
 - tap g key
 - let go of everything

Each time he will move to the "next" mode, when at the last mode he will
go back to the first.

Signed-off-by: Andy Green <andy.green@linaro.org>