adv7511 force hpd and fallback mode if no hpd at boot
If the HDMI connector is out during boot, this forces the HDMI up anyway
and causes it to select the fallback (720p) mode since no EDID is coming.

It means the HDMI is configured and working for 720p all the time.

This provides a neat way to force the fallback 720p mode by unplugging
until boot is complete, then plugging the monitor to find whatever booted
(Xorg, framebuffer console) configured fully for the fallback mode.

Signed-off-by: Andy Green <andy.green@linaro.org>