mmc: dw_mmc: abort command if card busy
mmc_rescan must fail if the card is busy (tested on dw_mmc-k3,
hisi6220 SoC).

This is an extension of Doug's Andersons previous commit:

Author: Doug Anderson <dianders@chromium.org>
Date:   Fri Feb 20 12:31:56 2015 -0800

    mmc: dw_mmc: Don't start commands while busy

    Signed-off-by: Doug Anderson <dianders@chromium.org>
    Tested-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>
    Signed-off-by: Jaehoon Chung <jh80.chung@samsung.com>
    Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>

Ignoring this condition on hisilicon's 6220 Soc causes HLE interrupts
to be raised.

Since theinterrupt  handler routine never services this interrupt this
ends up causing a never ending storm.

In order to be able to recover we also need to make sure that the card
is not busy right after a power up. So check for that condition in the
core.

Signed-off-by: Jorge Ramirez-Ortiz <jorge.ramirez-ortiz@linaro.org>